Create AppCompositionRoot to provide global service level app
