+++ app/src/main/java/com/techyourchance/dagger2course/common/composition/AppCompositionRoot.kt
@@ -1,0 +1,19 @@
+package com.techyourchance.dagger2course.common.composition
+
+import com.techyourchance.dagger2course.Constants
+import com.techyourchance.dagger2course.data.usecase.FetchQuestionsUseCaseImpl
+import com.techyourchance.dagger2course.networking.StackoverflowApi
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+class AppCompositionRoot {
+
+    private val retrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(Constants.BASE_URL)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    private val stackoverflowApi: StackoverflowApi = retrofit.create(StackoverflowApi::class.java)
+
+    val fetchQuestionsUseCase get() = FetchQuestionsUseCaseImpl(stackoverflowApi)
+}
