+++ app/src/main/java/com/techyourchance/dagger2course/screens/common/activities/BaseActivity.kt
@@ -1,0 +1,11 @@
+package com.techyourchance.dagger2course.screens.common.activities
+
+import androidx.appcompat.app.AppCompatActivity
+import com.techyourchance.dagger2course.MyApplication
+import com.techyourchance.dagger2course.common.composition.AppCompositionRoot
+
+open class BaseActivity: AppCompatActivity() {
+
+    protected val compositionRoot: AppCompositionRoot get() = (application as MyApplication).appCompositionRoot
+
+}
